+++ src/main/java/pers/solid/brrp/v1/tag/ObjectTagBuilder.java
@@ -56,7 +56,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
   }
 
   @Override
-  @Contract(mutates = "this")
   public Self add(@NotNull TagEntry entry) {
     super.add(entry);
     return self();
@@ -68,7 +67,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    * @param id The id of the entry to add. It is not a tag id.
    */
   @Override
-  @Contract(mutates = "this")
   public Self add(@NotNull Identifier id) {
     super.add(id);
     return self();
@@ -80,7 +78,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    * @param registryKey The registry key of the entry to be added.
    */
   @ApiStatus.AvailableSince("1.1.0")
-  @Contract(mutates = "this")
   public Self add(@NotNull RegistryKey<T> registryKey) {
     return this.add(registryKey.getValue());
   }
@@ -91,7 +88,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    * @param id The id of the entry to add. It is not a tag id.
    */
   @Override
-  @Contract(mutates = "this")
   public Self addOptional(@NotNull Identifier id) {
     super.addOptional(id);
     return self();
@@ -103,7 +99,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    * @param registryKey The registry key of the entry to add. It is not a tag id.
    */
   @ApiStatus.AvailableSince("1.1.0")
-  @Contract(mutates = "this")
   public Self addOptional(@NotNull RegistryKey<T> registryKey) {
     return addOptional(registryKey.getValue());
   }
@@ -114,7 +109,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    * @param id The id of the tag to be added to this tag.
    */
   @Override
-  @Contract(mutates = "this")
   public Self addTag(@NotNull Identifier id) {
     super.addTag(id);
     return self();
@@ -125,7 +119,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    *
    * @param tagKey The other tag to be added to this tag.
    */
-  @Contract(mutates = "this")
   public Self addTag(@NotNull TagKey<T> tagKey) {
     return addTag(tagKey.id());
   }
@@ -136,7 +129,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    * @param tagKeys The other tags to be added to this tag.
    */
   @SafeVarargs
-  @Contract(mutates = "this")
   public final Self addTag(@NotNull TagKey<T>... tagKeys) {
     for (TagKey<T> tagKey : tagKeys) {
       this.addTag(tagKey);
@@ -144,12 +136,10 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
     return self();
   }
 
-  @Contract(mutates = "this")
   public Self addTag(@NotNull IdentifiedTagBuilder<T> tagBuilder) {
     return addTag(tagBuilder.identifier);
   }
 
-  @Contract(mutates = "this")
   @SafeVarargs
   public final Self addTag(@NotNull IdentifiedTagBuilder<T>... tagBuilders) {
     for (IdentifiedTagBuilder<T> tagBuilder : tagBuilders) {
@@ -163,7 +153,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    *
    * @param id The id of the tag to be added to this tag.
    */
-  @Contract(mutates = "this")
   @Override
   public Self addOptionalTag(@NotNull Identifier id) {
     super.addOptionalTag(id);
@@ -175,7 +164,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    *
    * @param value The object to be added to this tag.
    */
-  @Contract(mutates = "this")
   public Self add(T value) {
     return add(valueToId.apply(value));
   }
@@ -185,7 +173,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    *
    * @param value The objects to be added to this tag.
    */
-  @Contract(mutates = "this")
   @SafeVarargs
   public final Self add(T... value) {
     Arrays.stream(value).map(valueToId).forEach(this::add);
@@ -198,7 +185,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    * @param identifiers The identifiers to be added to this tag.
    */
   @ApiStatus.AvailableSince("1.1.0")
-  @Contract(mutates = "this")
   public final Self add(Identifier... identifiers) {
     for (Identifier identifier : identifiers) {
       add(identifier);
@@ -213,7 +199,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
    */
   @SafeVarargs
   @ApiStatus.AvailableSince("1.1.0")
-  @Contract(mutates = "this")
   public final Self add(RegistryKey<T>... registryKeys) {
     for (RegistryKey<T> tagKey : registryKeys) {
       add(tagKey);
@@ -221,7 +206,6 @@ public class ObjectTagBuilder<T, Self extends ObjectTagBuilder<T, Self>> extends
     return self();
   }
 
-  @Contract(mutates = "this")
   public Self copy(TagBuilder copyFrom) {
     ((TagBuilderAccessor) copyFrom).getEntries().forEach(this::add);
     return self();